pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "calendar-app-mcp"
version = "0.1.0"
description = "Calendar app MCP for accessing macOS Calendar events and reminders"
requires-python = ">= 3.12"
dependencies = [
    "pyobjc-framework-EventKit",
    "fastmcp",
    "jinja2>=3.1.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.11.0",
    "black>=23.0.0",
    "pytest-black>=0.3.0",
    "pytest-ruff>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["calendar_app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
markers = [
    "unit: marks tests as unit tests",
    "integration: marks tests as integration tests", 
    "black: run the black formatter on all Python files",
    "ruff: run the ruff linter on Python files",
]

[tool.coverage.run]
source = ["calendar_app"]
omit = ["*/__init__.py", "*/__main__.py"]

# This makes the package accessible via `uv run`
[project.scripts]
calendar-app = "calendar_app.cli:main"

# Override the extra-index-url from the global config
[tool.uv]
index-url = "https://pypi.org/simple"
extra-index-url = []
index-strategy = "unsafe-best-match"

[tool.black]
line-length = 100
target-version = ["py312"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "C4", "EM", "Q", "T20", "UP", "ANN", "FLY", "LOG", "RUF", "SIM"]
ignore = [
    "ANN001", "ANN101", "ANN102", "ANN201", "ANN202", "ANN204", "ANN401",  # Allow missing type annotations
    "T201",  # Allow print statements
    "B904",  # Allow raise without from
    "EM101", "EM102",  # Allow string literals in exceptions
    "E501",  # Allow long lines
    "RUF013"  # Allow implicit Optional
]

[tool.ruff.lint.isort]
known-first-party = ["calendar_app"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ANN", "S", "T", "E", "W", "F", "I", "SIM", "F4", "F8"]

